pyproject.toml
[project]
name = "ibl-aws"
dynamic = [ "version" ]
description = "Default template for PDM package"
authors = [
    {name = "Olivier Winter", email = "olivier.winter@internationalbrainlab.org"},
]
dependencies = [
    "boto3",
    "python-dotenv",
    "paramiko",
    "iblutil",
    "lightning_sdk",
    "pydantic",
    "pytest",
    "pytest-mock",
]

requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.version]
source = "file"
path = 'src/iblaws/__init__.py'

[tool.pdm.dev-dependencies]
test = [
    "ruff",
]

[tool.pdm.build]
package-dir = "src"

[tool.pdm]
distribution = true

[tool.ruff]
exclude = [
    ".venv",
]
line-length = 130

[tool.ruff.format]
quote-style = "single"
